+++ core/src/network/peer.rs
@@ -35,8 +35,11 @@ use crate::{
         IntoConnectionHandler,
         PendingConnection,
         Substream,
+        pool::Pool,
     },
 };
+use fnv::FnvHashMap;
+use smallvec::SmallVec;
 use std::{
     collections::hash_map,
     error,
@@ -47,6 +50,10 @@ use super::{Network, DialingOpts};
 
 /// The possible representations of a peer in a [`Network`], as
 /// seen by the local node.
+///
+/// > **Note**: In any state there may always be a pending incoming
+/// > connection attempt from the peer, however, the remote identity
+/// > of a peer is only known once a connection is fully established.
 pub enum Peer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
 where
     TTrans: Transport,
@@ -63,10 +70,6 @@ where
     /// There exists no established connection to the peer and there is
     /// currently no ongoing dialing (i.e. outgoing connection) attempt
     /// in progress.
-    ///
-    /// > **Note**: In this state there may always be a pending incoming
-    /// > connection attempt from the peer, however, the remote identity
-    /// > of a peer is only known once a connection is fully established.
     Disconnected(DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>),
 
     /// The peer represents the local node.
@@ -82,20 +85,20 @@ where
     TPeerId: fmt::Debug + Eq + Hash,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        match *self {
-            Peer::Connected(ConnectedPeer { ref peer_id, .. }) => {
+        match self {
+            Peer::Connected(p) => {
                 f.debug_struct("Connected")
-                    .field("peer_id", peer_id)
+                    .field("peer", &p)
                     .finish()
             }
-            Peer::Dialing(DialingPeer { ref peer_id, .. } ) => {
-                f.debug_struct("DialingPeer")
-                    .field("peer_id", peer_id)
+            Peer::Dialing(p) => {
+                f.debug_struct("Dialing")
+                    .field("peer", &p)
                     .finish()
             }
-            Peer::Disconnected(DisconnectedPeer { ref peer_id, .. }) => {
+            Peer::Disconnected(p) => {
                 f.debug_struct("Disconnected")
-                    .field("peer_id", peer_id)
+                    .field("peer", &p)
                     .finish()
             }
             Peer::Local => {
@@ -164,12 +167,11 @@ where
     TTrans::Dial: Send + 'static,
     TMuxer: StreamMuxer + Send + Sync + 'static,
     TMuxer::OutboundSubstream: Send,
-    TMuxer::Substream: Send,
     TInEvent: Send + 'static,
     TOutEvent: Send + 'static,
     THandler: IntoConnectionHandler<TConnInfo> + Send + 'static,
-    THandler::Handler: ConnectionHandler<Substream = Substream<TMuxer>, InEvent = TInEvent, OutEvent = TOutEvent> + Send + 'static,
-    <THandler::Handler as ConnectionHandler>::OutboundOpenInfo: Send + 'static, // TODO: shouldn't be necessary
+    THandler::Handler: ConnectionHandler<Substream = Substream<TMuxer>, InEvent = TInEvent, OutEvent = TOutEvent> + Send,
+    <THandler::Handler as ConnectionHandler>::OutboundOpenInfo: Send,
     <THandler::Handler as ConnectionHandler>::Error: error::Error + Send + 'static,
     TConnInfo: fmt::Debug + ConnectionInfo<PeerId = TPeerId> + Send + 'static,
     TPeerId: Eq + Hash + Clone + Send + 'static,
@@ -208,7 +210,41 @@ where
         }
     }
 
-    /// Converts the peer into a `ConnectedPeer`, if there an established connection exists.
+    /// Initiates a new dialing attempt to this peer using the given addresses.
+    ///
+    /// The connection ID of the first connection attempt, i.e. to `address`,
+    /// is returned, together with a [`DialingPeer`] for further use. The
+    /// `remaining` addresses are tried in order in subsequent connection
+    /// attempts in the context of the same dialing attempt, if the connection
+    /// attempt to the first address fails.
+    pub fn dial<I>(self, address: Multiaddr, remaining: I, handler: THandler)
+        -> Result<
+            (ConnectionId, DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>),
+            ConnectionLimit
+        >
+    where
+        I: IntoIterator<Item = Multiaddr>,
+    {
+        let (peer_id, network) = match self {
+            Peer::Connected(p) => (p.peer_id, p.network),
+            Peer::Dialing(p) => (p.peer_id, p.network),
+            Peer::Disconnected(p) => (p.peer_id, p.network),
+            Peer::Local => return Err(ConnectionLimit { current: 0, limit: 0 })
+        };
+
+        let id = network.dial_peer(DialingOpts {
+            peer: peer_id.clone(),
+            handler,
+            address,
+            remaining: remaining.into_iter().collect(),
+        })?;
+
+        Ok((id, DialingPeer { network, peer_id }))
+    }
+
+    /// Converts the peer into a `ConnectedPeer`, if an established connection exists.
+    ///
+    /// Succeeds if the there is at least one established connection to the peer.
     pub fn into_connected(self) -> Option<
         ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
     > {
@@ -221,6 +257,8 @@ where
     }
 
     /// Converts the peer into a `DialingPeer`, if a dialing attempt exists.
+    ///
+    /// Succeeds if the there is at least one pending outgoing connection to the peer.
     pub fn into_dialing(self) -> Option<
         DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
     > {
@@ -245,7 +283,8 @@ where
 }
 
 /// The representation of a peer in a [`Network`] to whom at least
-/// one established connection exists.
+/// one established connection exists. There may also be additional ongoing
+/// dialing attempts to the peer.
 pub struct ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
 where
     TTrans: Transport,
@@ -267,57 +306,12 @@ where
         &self.peer_id
     }
 
-    /// Attempts to establish a new connection to this peer using the given addresses,
-    /// if there is currently no ongoing dialing attempt.
-    ///
-    /// Existing established connections are not affected.
-    ///
-    /// > **Note**: If there is an ongoing dialing attempt, a `DialingPeer`
-    /// > is returned with the given addresses and handler being ignored.
-    /// > You may want to check [`ConnectedPeer::is_dialing`] first.
-    pub fn connect<I, TMuxer>(self, address: Multiaddr, remaining: I, handler: THandler)
-        -> Result<DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>,
-                  ConnectionLimit>
-    where
-        I: IntoIterator<Item = Multiaddr>,
-        THandler: Send + 'static,
-        THandler::Handler: Send,
-        <THandler::Handler as ConnectionHandler>::Error: error::Error + Send,
-        <THandler::Handler as ConnectionHandler>::OutboundOpenInfo: Send,
-        THandler::Handler: ConnectionHandler<Substream = Substream<TMuxer>, InEvent = TInEvent, OutEvent = TOutEvent> + Send,
-        TTrans: Transport<Output = (TConnInfo, TMuxer)> + Clone,
-        TTrans::Error: Send + 'static,
-        TTrans::Dial: Send + 'static,
-        TMuxer: StreamMuxer + Send + Sync + 'static,
-        TMuxer::OutboundSubstream: Send,
-        TMuxer::Substream: Send,
-        TConnInfo: fmt::Debug + Send + 'static,
-        TPeerId: Eq + Hash + Clone + Send + 'static,
-        TInEvent: Send + 'static,
-        TOutEvent: Send + 'static,
-
-    {
-        if self.network.dialing.contains_key(&self.peer_id) {
-            let peer = DialingPeer {
-                network: self.network,
-                peer_id: self.peer_id
-            };
-            Ok(peer)
-        } else {
-           self.network.dial_peer(DialingOpts {
-                peer: self.peer_id.clone(),
-                handler,
-                address,
-                remaining: remaining.into_iter().collect(),
-            })?;
-            Ok(DialingPeer {
-                network: self.network,
-                peer_id: self.peer_id,
-            })
-        }
+    /// Returns the `ConnectedPeer` into a `Peer`.
+    pub fn into_peer(self) -> Peer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> {
+        Peer::Connected(self)
     }
 
-    /// Obtains an existing connection to the peer.
+    /// Obtains an established connection to the peer by ID.
     pub fn connection<'b>(&'b mut self, id: ConnectionId)
         -> Option<EstablishedConnection<'b, TInEvent, TConnInfo, TPeerId>>
     {
@@ -348,7 +342,7 @@ where
         }
     }
 
-    /// Gets an iterator over all established connections of the peer.
+    /// Gets an iterator over all established connections to the peer.
     pub fn connections<'b>(&'b mut self) ->
         EstablishedConnectionIter<'b,
             impl Iterator<Item = ConnectionId>,
@@ -386,11 +380,13 @@ impl<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> fmt::Debug f
 where
     TTrans: Transport,
     THandler: IntoConnectionHandler<TConnInfo>,
-    TPeerId: fmt::Debug,
+    TPeerId: Eq + Hash + fmt::Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         f.debug_struct("ConnectedPeer")
             .field("peer_id", &self.peer_id)
+            .field("established", &self.network.pool.iter_peer_established_info(&self.peer_id))
+            .field("attempts", &self.network.dialing.get(&self.peer_id))
             .finish()
     }
 }
@@ -419,8 +415,16 @@ where
         &self.peer_id
     }
 
-    /// Disconnects from this peer, closing all pending connections.
-    pub fn disconnect(self) -> DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> {
+    /// Returns the `DialingPeer` into a `Peer`.
+    pub fn into_peer(self) -> Peer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> {
+        Peer::Dialing(self)
+    }
+
+    /// Disconnects from this peer, closing all established connections and
+    /// aborting all dialing attempts.
+    pub fn disconnect(self)
+        -> DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
+    {
         self.network.disconnect(&self.peer_id);
         DisconnectedPeer { network: self.network, peer_id: self.peer_id }
     }
@@ -443,20 +447,50 @@ where
         }
     }
 
-    /// Obtains the connection that is currently being established.
-    pub fn connection<'b>(&'b mut self) -> DialingConnection<'b, TInEvent, TConnInfo, TPeerId> {
-        let attempt = match self.network.dialing.entry(self.peer_id.clone()) {
-            hash_map::Entry::Occupied(e) => e,
-            _ => unreachable!("By `Peer::new` and the definition of `DialingPeer`.")
-        };
-
-        let inner = self.network.pool
-            .get_outgoing(attempt.get().id)
-            .expect("By consistency of `network.pool` with `network.dialing`.");
-
-        DialingConnection {
-            inner, dialing: attempt, peer_id: &self.peer_id
+    /// Obtains a dialing attempt to the peer by connection ID of
+    /// the current connection attempt.
+    pub fn attempt<'b>(&'b mut self, id: ConnectionId)
+        -> Option<DialingAttempt<'b, TInEvent, TConnInfo, TPeerId>>
+    {
+        if let hash_map::Entry::Occupied(attempts) = self.network.dialing.entry(self.peer_id.clone()) {
+            if let Some(pos) = attempts.get().iter().position(|s| s.current.0 == id) {
+                if let Some(inner) = self.network.pool.get_outgoing(id) {
+                    return Some(DialingAttempt { pos, inner, attempts })
+                }
+            }
         }
+        None
+    }
+
+    /// The number of ongoing dialing attempts, i.e. pending outgoing connections
+    /// to this peer.
+    pub fn num_attempts(&self) -> usize {
+        self.network.pool.num_peer_outgoing(&self.peer_id)
+    }
+
+    /// Gets an iterator over all dialing (i.e. pending outgoing) connections to the peer.
+    pub fn attempts<'b>(&'b mut self)
+        -> DialingAttemptIter<'b,
+            TInEvent,
+            TOutEvent,
+            THandler,
+            TTrans::Error,
+            <THandler::Handler as ConnectionHandler>::Error,
+            TConnInfo,
+            TPeerId>
+    {
+        DialingAttemptIter::new(&self.peer_id, &mut self.network.pool, &mut self.network.dialing)
+    }
+
+    /// Obtains some dialing connection to the peer.
+    ///
+    /// At least one dialing connection is guaranteed to exist on a `DialingPeer`.
+    pub fn some_attempt<'b>(&'b mut self)
+        -> DialingAttempt<'b, TInEvent, TConnInfo, TPeerId>
+    {
+        self.attempts()
+            .into_first()
+            .expect("By `Peer::new` and the definition of `DialingPeer`.")
     }
 }
 
@@ -465,11 +499,13 @@ impl<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> fmt::Debug f
 where
     TTrans: Transport,
     THandler: IntoConnectionHandler<TConnInfo>,
-    TPeerId: fmt::Debug,
+    TPeerId: Eq + Hash + fmt::Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         f.debug_struct("DialingPeer")
             .field("peer_id", &self.peer_id)
+            .field("established", &self.network.pool.iter_peer_established_info(&self.peer_id))
+            .field("attempts", &self.network.dialing.get(&self.peer_id))
             .finish()
     }
 }
@@ -500,46 +536,19 @@ where
     }
 }
 
-impl<'a, TTrans, TInEvent, TOutEvent, TMuxer, THandler, TConnInfo, TPeerId>
+impl<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
     DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>
 where
-    TTrans: Transport<Output = (TConnInfo, TMuxer)> + Clone,
-    TTrans::Error: Send + 'static,
-    TTrans::Dial: Send + 'static,
-    TMuxer: StreamMuxer + Send + Sync + 'static,
-    TMuxer::OutboundSubstream: Send,
-    TMuxer::Substream: Send,
-    THandler: IntoConnectionHandler<TConnInfo> + Send + 'static,
-    THandler::Handler: ConnectionHandler<Substream = Substream<TMuxer>, InEvent = TInEvent, OutEvent = TOutEvent> + Send,
-    <THandler::Handler as ConnectionHandler>::OutboundOpenInfo: Send,
-    <THandler::Handler as ConnectionHandler>::Error: error::Error + Send,
-    TInEvent: Send + 'static,
-    TOutEvent: Send + 'static,
+    TTrans: Transport,
+    THandler: IntoConnectionHandler<TConnInfo>,
 {
     pub fn id(&self) -> &TPeerId {
         &self.peer_id
     }
 
-    /// Attempts to connect to this peer using the given addresses.
-    pub fn connect<TIter>(self, first: Multiaddr, rest: TIter, handler: THandler)
-        -> Result<DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>,
-                  ConnectionLimit>
-    where
-        TIter: IntoIterator<Item = Multiaddr>,
-        TConnInfo: fmt::Debug + ConnectionInfo<PeerId = TPeerId> + Send + 'static,
-        TPeerId: Eq + Hash + Clone + Send + 'static,
-    {
-        self.network.dial_peer(DialingOpts {
-            peer: self.peer_id.clone(),
-            handler,
-            address: first,
-            remaining: rest.into_iter().collect(),
-        })?;
-        Ok(DialingPeer {
-            network: self.network,
-            peer_id: self.peer_id,
-        })
-
+    /// Returns the `DisconnectedPeer` into a `Peer`.
+    pub fn into_peer(self) -> Peer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> {
+        Peer::Disconnected(self)
     }
 
     /// Moves the peer into a connected state by supplying an existing
@@ -550,8 +559,7 @@ where
     /// # Panics
     ///
     /// Panics if `connected.peer_id()` does not identify the current peer.
-    ///
-    pub fn set_connected(
+    pub fn set_connected<TMuxer>(
         self,
         connected: Connected<TConnInfo>,
         connection: Connection<TMuxer, THandler::Handler>,
@@ -559,8 +567,17 @@ where
         ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>,
         ConnectionLimit
     > where
+        TInEvent: Send + 'static,
+        TOutEvent: Send + 'static,
+        THandler: Send + 'static,
+        TTrans::Error: Send + 'static,
+        THandler::Handler: ConnectionHandler<Substream = Substream<TMuxer>, InEvent = TInEvent, OutEvent = TOutEvent> + Send,
+        <THandler::Handler as ConnectionHandler>::OutboundOpenInfo: Send,
+        <THandler::Handler as ConnectionHandler>::Error: error::Error + Send + 'static,
         TConnInfo: fmt::Debug + ConnectionInfo<PeerId = TPeerId> + Clone + Send + 'static,
-        TPeerId: Eq + Hash + Clone + fmt::Debug,
+        TPeerId: Eq + Hash + Clone + Send + fmt::Debug + 'static,
+        TMuxer: StreamMuxer + Send + Sync + 'static,
+        TMuxer::OutboundSubstream: Send,
     {
         if connected.peer_id() != &self.peer_id {
             panic!("Invalid peer ID given: {:?}. Expected: {:?}", connected.peer_id(), self.peer_id)
@@ -574,71 +591,142 @@ where
     }
 }
 
-/// Attempt to reach a peer.
+/// The (internal) state of a `DialingAttempt`, tracking the
+/// current connection attempt as well as remaining addresses.
 #[derive(Debug, Clone)]
-pub(super) struct DialingAttempt {
-    /// Identifier for the reach attempt.
-    pub(super) id: ConnectionId,
-    /// Multiaddr currently being attempted.
-    pub(super) current: Multiaddr,
+pub(super) struct DialingState {
+    /// The ID and (remote) address of the current connection attempt.
+    pub(super) current: (ConnectionId, Multiaddr),
     /// Multiaddresses to attempt if the current one fails.
-    pub(super) next: Vec<Multiaddr>,
+    pub(super) remaining: Vec<Multiaddr>,
 }
 
-/// A `DialingConnection` is a [`PendingConnection`] where the local peer
-/// has the role of the dialer (i.e. initiator) and the (expected) remote
-/// peer ID is known.
-pub struct DialingConnection<'a, TInEvent, TConnInfo, TPeerId> {
-    peer_id: &'a TPeerId,
+/// A `DialingAttempt` is an ongoing outgoing connection attempt to
+/// a known / expected remote peer ID and a list of alternative addresses
+/// to connect to, if the current connection attempt fails.
+pub struct DialingAttempt<'a, TInEvent, TConnInfo, TPeerId> {
+    /// The underlying pending connection in the `Pool`.
     inner: PendingConnection<'a, TInEvent, TConnInfo, TPeerId>,
-    dialing: hash_map::OccupiedEntry<'a, TPeerId, DialingAttempt>,
+    /// All current dialing attempts of the peer.
+    attempts: hash_map::OccupiedEntry<'a, TPeerId, SmallVec<[DialingState; 10]>>,
+    /// The position of the current `DialingState` of this connection in the `attempts`.
+    pos: usize,
 }
 
 impl<'a, TInEvent, TConnInfo, TPeerId>
-    DialingConnection<'a, TInEvent, TConnInfo, TPeerId>
+    DialingAttempt<'a, TInEvent, TConnInfo, TPeerId>
 {
-    /// Returns the local connection ID.
+    /// Returns the ID of the current connection attempt.
     pub fn id(&self) -> ConnectionId {
         self.inner.id()
     }
 
-    /// Returns the (expected) peer ID of the ongoing connection attempt.
+    /// Returns the (expected) peer ID of the dialing attempt.
     pub fn peer_id(&self) -> &TPeerId {
-        self.peer_id
+        self.attempts.key()
     }
 
-    /// Returns information about this endpoint of the connection attempt.
-    pub fn endpoint(&self) -> &ConnectedPoint {
-        self.inner.endpoint()
-    }
-
-    /// Aborts the connection attempt.
-    pub fn abort(self)
-    where
-        TPeerId: Eq + Hash + Clone,
-    {
-        self.dialing.remove();
-        self.inner.abort();
-    }
-
-    /// Adds new candidate addresses to the end of the addresses used
-    /// in the ongoing dialing process.
-    ///
-    /// Duplicates are ignored.
-    pub fn add_addresses(&mut self, addrs: impl IntoIterator<Item = Multiaddr>) {
-        for addr in addrs {
-            self.add_address(addr);
+    /// Returns the remote address of the current connection attempt.
+    pub fn address(&self) -> &Multiaddr {
+        match self.inner.endpoint() {
+            ConnectedPoint::Dialer { address } => address,
+            ConnectedPoint::Listener { .. } => unreachable!("by definition of a `DialingAttempt`.")
         }
     }
 
-    /// Adds an address to the end of the addresses used in the ongoing
-    /// dialing process.
+    /// Aborts the dialing attempt.
     ///
-    /// Duplicates are ignored.
+    /// Aborting a dialing attempt involves aborting the current connection
+    /// attempt and dropping any remaining addresses given to [`Peer::dial()`]
+    /// that have not yet been tried.
+    pub fn abort(mut self) {
+        self.attempts.get_mut().remove(self.pos);
+        if self.attempts.get().is_empty() {
+            self.attempts.remove();
+        }
+        self.inner.abort();
+    }
+
+    /// Adds an address to the end of the remaining addresses
+    /// for this dialing attempt. Duplicates are ignored.
     pub fn add_address(&mut self, addr: Multiaddr) {
-        if self.dialing.get().next.iter().all(|a| a != &addr) {
-            self.dialing.get_mut().next.push(addr);
+        let remaining = &mut self.attempts.get_mut()[self.pos].remaining;
+        if remaining.iter().all(|a| a != &addr) {
+            remaining.push(addr);
         }
     }
 }
 
+/// An iterator over the ongoing dialing attempts to a peer.
+pub struct DialingAttemptIter<'a, TInEvent, TOutEvent, THandler, TTransErr, THandlerErr, TConnInfo, TPeerId> {
+    /// The peer whose dialing attempts are being iterated.
+    peer_id: &'a TPeerId,
+    /// The underlying connection `Pool` of the `Network`.
+    pool: &'a mut Pool<TInEvent, TOutEvent, THandler, TTransErr, THandlerErr, TConnInfo, TPeerId>,
+    /// The state of all current dialing attempts known to the `Network`.
+    ///
+    /// Ownership of the `OccupiedEntry` for `peer_id` containing all attempts must be
+    /// borrowed to each `DialingAttempt` in order for it to remove the entry if the
+    /// last dialing attempt is aborted.
+    dialing: &'a mut FnvHashMap<TPeerId, SmallVec<[DialingState; 10]>>,
+    /// The current position of the iterator in `dialing[peer_id]`.
+    pos: usize,
+    /// The total number of elements in `dialing[peer_id]` to iterate over.
+    end: usize,
+}
+
+// Note: Ideally this would be an implementation of `Iterator`, but that
+// requires GATs (cf. https://github.com/rust-lang/rust/issues/44265) and
+// a different definition of `Iterator`.
+impl<'a, TInEvent, TOutEvent, THandler, TTransErr, THandlerErr, TConnInfo, TPeerId>
+    DialingAttemptIter<'a, TInEvent, TOutEvent, THandler, TTransErr, THandlerErr, TConnInfo, TPeerId>
+where
+    TConnInfo: ConnectionInfo<PeerId = TPeerId>,
+    TPeerId: Eq + Hash + Clone,
+{
+    fn new(
+        peer_id: &'a TPeerId,
+        pool: &'a mut Pool<TInEvent, TOutEvent, THandler, TTransErr, THandlerErr, TConnInfo, TPeerId>,
+        dialing: &'a mut FnvHashMap<TPeerId, SmallVec<[DialingState; 10]>>,
+    ) -> Self {
+        let end = dialing.get(peer_id).map_or(0, |conns| conns.len());
+        Self { pos: 0, end, pool, dialing, peer_id }
+    }
+
+    /// Obtains the next dialing connection, if any.
+    pub fn next<'b>(&'b mut self) -> Option<DialingAttempt<'b, TInEvent, TConnInfo, TPeerId>> {
+        if self.pos == self.end {
+            return None
+        }
+
+        if let hash_map::Entry::Occupied(attempts) = self.dialing.entry(self.peer_id.clone()) {
+            let id = attempts.get()[self.pos].current.0;
+            if let Some(inner) = self.pool.get_outgoing(id) {
+                let conn = DialingAttempt { pos: self.pos, inner, attempts };
+                self.pos += 1;
+                return Some(conn)
+            }
+        }
+
+        None
+    }
+
+    /// Returns the first connection, if any, consuming the iterator.
+    pub fn into_first<'b>(self)
+        -> Option<DialingAttempt<'b, TInEvent, TConnInfo, TPeerId>>
+    where 'a: 'b
+    {
+        if self.pos == self.end {
+            return None
+        }
+
+        if let hash_map::Entry::Occupied(attempts) = self.dialing.entry(self.peer_id.clone()) {
+            let id = attempts.get()[self.pos].current.0;
+            if let Some(inner) = self.pool.get_outgoing(id) {
+                return Some(DialingAttempt { pos: self.pos, inner, attempts })
+            }
+        }
+
+        None
+    }
+}